Other files moved to utils folder
